Added myself as a contributor
